Summarize a Workflow-Layer Result

Summarize the output of run_testing_workflow into a compact reporting object. This function belongs to the reporting layer; it does not construct matrices, simulate outcomes, or run decoders.

Usage

protocol_summary(x)

Arguments

x

A workflow result object returned by run_testing_workflow.

Value

A named list containing compact workflow counts, including the number of pools, the number of individuals, the number of stage-1 non-negative calls, the number of planned stage-2 retests, and the counts of final positive, negative, and unresolved individuals.

Details

Let \(\mathbf{M} \in \{0,1\}^{J \times N}\) denote the stage-1 pooling design matrix and let \(\mathbf{z}\) denote the stage-1 pool outcomes. A workflow result records stage-1 calls, optional stage-2 retests, and final consolidated calls. The summary returned by protocol_summary() is a reporting view of those stored objects. It does not re-run any protocol logic.

See also

Examples

M <- pp_matrix(q = 4, d = 3, nl = 5)
M
#> 20 x 64 sparse Matrix of class "dgCMatrix"
#>                                                                                
#>  [1,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#>  [2,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#>  [3,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#>  [4,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>  [5,] 1 . . . . 1 . . . . 1 . . . . 1 . 1 . . 1 . . . . . . 1 . . 1 . . . 1 . .
#>  [6,] . 1 . . 1 . . . . . . 1 . . 1 . 1 . . . . 1 . . . . 1 . . . . 1 . . . 1 .
#>  [7,] . . 1 . . . . 1 1 . . . . 1 . . . . . 1 . . 1 . . 1 . . 1 . . . 1 . . . .
#>  [8,] . . . 1 . . 1 . . 1 . . 1 . . . . . 1 . . . . 1 1 . . . . 1 . . . 1 . . 1
#>  [9,] 1 . . . . . 1 . . . . 1 . 1 . . . . . 1 . 1 . . 1 . . . . . 1 . . 1 . . .
#> [10,] . 1 . . . . . 1 . . 1 . 1 . . . . . 1 . 1 . . . . 1 . . . . . 1 1 . . . .
#> [11,] . . 1 . 1 . . . . 1 . . . . . 1 . 1 . . . . . 1 . . 1 . 1 . . . . . . 1 .
#> [12,] . . . 1 . 1 . . 1 . . . . . 1 . 1 . . . . . 1 . . . . 1 . 1 . . . . 1 . 1
#> [13,] 1 . . . . . . 1 . 1 . . . . 1 . . . 1 . . 1 . . . . . 1 1 . . . . . . 1 1
#> [14,] . 1 . . . . 1 . 1 . . . . . . 1 . . . 1 1 . . . . . 1 . . 1 . . . . 1 . .
#> [15,] . . 1 . . 1 . . . . . 1 1 . . . 1 . . . . . . 1 . 1 . . . . 1 . . 1 . . .
#> [16,] . . . 1 1 . . . . . 1 . . 1 . . . 1 . . . . 1 . 1 . . . . . . 1 1 . . . .
#> [17,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#> [18,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#> [19,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#> [20,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>                                                            
#>  [1,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#>  [2,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#>  [3,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>  [4,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#>  [5,] . . 1 1 . . . . 1 . . . . . 1 . . 1 . . 1 . . 1 . . .
#>  [6,] . 1 . . 1 . . 1 . . . . . 1 . . . . 1 1 . . . . 1 . .
#>  [7,] 1 . . . . 1 . . . . 1 . 1 . . 1 . . . . . . 1 . . 1 .
#>  [8,] . . . . . . 1 . . 1 . 1 . . . . 1 . . . . 1 . . . . 1
#>  [9,] . . 1 . . 1 . 1 . . . . . 1 . 1 . . . . 1 . . . . . 1
#> [10,] . 1 . . . . 1 . 1 . . . . . 1 . 1 . . 1 . . . . . 1 .
#> [11,] 1 . . 1 . . . . . 1 . 1 . . . . . 1 . . . . 1 . 1 . .
#> [12,] . . . . 1 . . . . . 1 . 1 . . . . . 1 . . 1 . 1 . . .
#> [13,] . . . . . 1 . . 1 . . . 1 . . . . 1 . 1 . . . . . . 1
#> [14,] 1 . . . . . 1 1 . . . 1 . . . . . . 1 . 1 . . . . 1 .
#> [15,] . 1 . 1 . . . . . . 1 . . . 1 1 . . . . . 1 . . 1 . .
#> [16,] . . 1 . 1 . . . . 1 . . . 1 . . 1 . . . . . 1 1 . . .
#> [17,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#> [18,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#> [19,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#> [20,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1

Y_tilde <- rep(0L, 64L)
Y_tilde[c(2L, 13L)] <- 1L
Y_tilde
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [39] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0

sim <- simulate_group_testing(M, Y_tilde)
sim
#> $matrix
#> 20 x 64 sparse Matrix of class "dgCMatrix"
#>                                                                                
#>  [1,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#>  [2,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#>  [3,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#>  [4,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>  [5,] 1 . . . . 1 . . . . 1 . . . . 1 . 1 . . 1 . . . . . . 1 . . 1 . . . 1 . .
#>  [6,] . 1 . . 1 . . . . . . 1 . . 1 . 1 . . . . 1 . . . . 1 . . . . 1 . . . 1 .
#>  [7,] . . 1 . . . . 1 1 . . . . 1 . . . . . 1 . . 1 . . 1 . . 1 . . . 1 . . . .
#>  [8,] . . . 1 . . 1 . . 1 . . 1 . . . . . 1 . . . . 1 1 . . . . 1 . . . 1 . . 1
#>  [9,] 1 . . . . . 1 . . . . 1 . 1 . . . . . 1 . 1 . . 1 . . . . . 1 . . 1 . . .
#> [10,] . 1 . . . . . 1 . . 1 . 1 . . . . . 1 . 1 . . . . 1 . . . . . 1 1 . . . .
#> [11,] . . 1 . 1 . . . . 1 . . . . . 1 . 1 . . . . . 1 . . 1 . 1 . . . . . . 1 .
#> [12,] . . . 1 . 1 . . 1 . . . . . 1 . 1 . . . . . 1 . . . . 1 . 1 . . . . 1 . 1
#> [13,] 1 . . . . . . 1 . 1 . . . . 1 . . . 1 . . 1 . . . . . 1 1 . . . . . . 1 1
#> [14,] . 1 . . . . 1 . 1 . . . . . . 1 . . . 1 1 . . . . . 1 . . 1 . . . . 1 . .
#> [15,] . . 1 . . 1 . . . . . 1 1 . . . 1 . . . . . . 1 . 1 . . . . 1 . . 1 . . .
#> [16,] . . . 1 1 . . . . . 1 . . 1 . . . 1 . . . . 1 . 1 . . . . . . 1 1 . . . .
#> [17,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#> [18,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#> [19,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#> [20,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>                                                            
#>  [1,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#>  [2,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#>  [3,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>  [4,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#>  [5,] . . 1 1 . . . . 1 . . . . . 1 . . 1 . . 1 . . 1 . . .
#>  [6,] . 1 . . 1 . . 1 . . . . . 1 . . . . 1 1 . . . . 1 . .
#>  [7,] 1 . . . . 1 . . . . 1 . 1 . . 1 . . . . . . 1 . . 1 .
#>  [8,] . . . . . . 1 . . 1 . 1 . . . . 1 . . . . 1 . . . . 1
#>  [9,] . . 1 . . 1 . 1 . . . . . 1 . 1 . . . . 1 . . . . . 1
#> [10,] . 1 . . . . 1 . 1 . . . . . 1 . 1 . . 1 . . . . . 1 .
#> [11,] 1 . . 1 . . . . . 1 . 1 . . . . . 1 . . . . 1 . 1 . .
#> [12,] . . . . 1 . . . . . 1 . 1 . . . . . 1 . . 1 . 1 . . .
#> [13,] . . . . . 1 . . 1 . . . 1 . . . . 1 . 1 . . . . . . 1
#> [14,] 1 . . . . . 1 1 . . . 1 . . . . . . 1 . 1 . . . . 1 .
#> [15,] . 1 . 1 . . . . . . 1 . . . 1 1 . . . . . 1 . . 1 . .
#> [16,] . . 1 . 1 . . . . 1 . . . 1 . . 1 . . . . . 1 1 . . .
#> [17,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#> [18,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#> [19,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#> [20,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#> 
#> $Y_tilde
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [39] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $z_tilde
#>  [1] 1 1 0 0 0 1 0 1 0 1 0 0 0 1 1 0 1 1 0 0
#> 
#> $z
#>  [1] 1 1 0 0 0 1 0 1 0 1 0 0 0 1 1 0 1 1 0 0
#> 
#> $s_e
#> NULL
#> 
#> $s_p
#> NULL
#> 
#> $mode
#> [1] "noiseless"
#> 

wf <- run_testing_workflow(
  design = M,
  z = sim$z,
  workflow = "non_adaptive",
  decoder = "pp_decode",
  decoder_args = list(verbose = FALSE)
)
wf
#> $workflow
#> [1] "non_adaptive"
#> 
#> $matrix
#> 20 x 64 sparse Matrix of class "dgCMatrix"
#>                                                                                
#>  [1,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#>  [2,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#>  [3,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#>  [4,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>  [5,] 1 . . . . 1 . . . . 1 . . . . 1 . 1 . . 1 . . . . . . 1 . . 1 . . . 1 . .
#>  [6,] . 1 . . 1 . . . . . . 1 . . 1 . 1 . . . . 1 . . . . 1 . . . . 1 . . . 1 .
#>  [7,] . . 1 . . . . 1 1 . . . . 1 . . . . . 1 . . 1 . . 1 . . 1 . . . 1 . . . .
#>  [8,] . . . 1 . . 1 . . 1 . . 1 . . . . . 1 . . . . 1 1 . . . . 1 . . . 1 . . 1
#>  [9,] 1 . . . . . 1 . . . . 1 . 1 . . . . . 1 . 1 . . 1 . . . . . 1 . . 1 . . .
#> [10,] . 1 . . . . . 1 . . 1 . 1 . . . . . 1 . 1 . . . . 1 . . . . . 1 1 . . . .
#> [11,] . . 1 . 1 . . . . 1 . . . . . 1 . 1 . . . . . 1 . . 1 . 1 . . . . . . 1 .
#> [12,] . . . 1 . 1 . . 1 . . . . . 1 . 1 . . . . . 1 . . . . 1 . 1 . . . . 1 . 1
#> [13,] 1 . . . . . . 1 . 1 . . . . 1 . . . 1 . . 1 . . . . . 1 1 . . . . . . 1 1
#> [14,] . 1 . . . . 1 . 1 . . . . . . 1 . . . 1 1 . . . . . 1 . . 1 . . . . 1 . .
#> [15,] . . 1 . . 1 . . . . . 1 1 . . . 1 . . . . . . 1 . 1 . . . . 1 . . 1 . . .
#> [16,] . . . 1 1 . . . . . 1 . . 1 . . . 1 . . . . 1 . 1 . . . . . . 1 1 . . . .
#> [17,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#> [18,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#> [19,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#> [20,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>                                                            
#>  [1,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#>  [2,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#>  [3,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#>  [4,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#>  [5,] . . 1 1 . . . . 1 . . . . . 1 . . 1 . . 1 . . 1 . . .
#>  [6,] . 1 . . 1 . . 1 . . . . . 1 . . . . 1 1 . . . . 1 . .
#>  [7,] 1 . . . . 1 . . . . 1 . 1 . . 1 . . . . . . 1 . . 1 .
#>  [8,] . . . . . . 1 . . 1 . 1 . . . . 1 . . . . 1 . . . . 1
#>  [9,] . . 1 . . 1 . 1 . . . . . 1 . 1 . . . . 1 . . . . . 1
#> [10,] . 1 . . . . 1 . 1 . . . . . 1 . 1 . . 1 . . . . . 1 .
#> [11,] 1 . . 1 . . . . . 1 . 1 . . . . . 1 . . . . 1 . 1 . .
#> [12,] . . . . 1 . . . . . 1 . 1 . . . . . 1 . . 1 . 1 . . .
#> [13,] . . . . . 1 . . 1 . . . 1 . . . . 1 . 1 . . . . . . 1
#> [14,] 1 . . . . . 1 1 . . . 1 . . . . . . 1 . 1 . . . . 1 .
#> [15,] . 1 . 1 . . . . . . 1 . . . 1 1 . . . . . 1 . . 1 . .
#> [16,] . . 1 . 1 . . . . 1 . . . 1 . . 1 . . . . . 1 1 . . .
#> [17,] . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . .
#> [18,] 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . .
#> [19,] . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 .
#> [20,] . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1 . . . 1
#> 
#> $z
#>  [1] 1 1 0 0 0 1 0 1 0 1 0 0 0 1 1 0 1 1 0 0
#> 
#> $stage1
#> $stage1$decoder
#> [1] "pp_decode"
#> 
#> $stage1$raw
#> $stage1$raw$positives
#> [1]  2 13
#> 
#> $stage1$raw$n_positives
#> [1] 2
#> 
#> $stage1$raw$suspected
#> integer(0)
#> 
#> $stage1$raw$n_suspected
#> [1] 0
#> 
#> $stage1$raw$candidates
#> [1]  2 13
#> 
#> $stage1$raw$n_candidates
#> [1] 2
#> 
#> $stage1$raw$error
#> [1] 0
#> 
#> $stage1$raw$method
#> [1] "COMP (exact)"
#> 
#> 
#> $stage1$positive
#> [1]  2 13
#> 
#> $stage1$suspected
#> integer(0)
#> 
#> $stage1$negative
#> integer(0)
#> 
#> $stage1$candidates
#> [1]  2 13
#> 
#> 
#> $stage2
#> $stage2$method
#> [1] "none"
#> 
#> $stage2$planned
#> integer(0)
#> 
#> $stage2$results
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA
#> [26] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA
#> [51] NA NA NA NA NA NA NA NA NA NA NA NA NA NA
#> 
#> $stage2$completed
#> [1] TRUE
#> 
#> 
#> $final_calls
#> $final_calls$positive
#> [1]  2 13
#> 
#> $final_calls$negative
#> integer(0)
#> 
#> $final_calls$unresolved
#>  [1]  1  3  4  5  6  7  8  9 10 11 12 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> [26] 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50 51 52
#> [51] 53 54 55 56 57 58 59 60 61 62 63 64
#> 
#> 
#> attr(,"class")
#> [1] "ppgt_workflow"

summary_obj <- protocol_summary(wf)
summary_obj
#> $workflow
#> [1] "non_adaptive"
#> 
#> $n_pools
#> [1] 20
#> 
#> $n_individuals
#> [1] 64
#> 
#> $n_stage1_positive
#> [1] 2
#> 
#> $n_stage1_suspected
#> [1] 0
#> 
#> $n_stage1_negative
#> [1] 0
#> 
#> $n_stage2_planned
#> [1] 0
#> 
#> $n_final_positive
#> [1] 2
#> 
#> $n_final_negative
#> [1] 0
#> 
#> $n_final_unresolved
#> [1] 62
#>